psweb/requirements.txt
django==2.0.1
django-debug-toolbar
postgres
psycopg2
django-crispy-forms==1.7.2
django-scheduler==0.8.8
django-money
django-model-utils==3.1.2
django_jchart
nltk
requests
owlready2
openpyxl
django-js-asset==1.2.2
django-mptt==0.10.0
django-extra-views==0.12.0
